bug fixes while introducing rosco

+++ wisdem/aeroelasticse/pyIECWind.py
@@ -206,7 +206,7 @@ class pyIECWind_extreme():
         self.setup()
         
         T = 10.
-        t = np.linspace(0., T, num=(T/self.dt+1))
+        t = np.linspace(0., T, num=int((T/self.dt+1)))
 
         # constants from standard
         alpha = 0.2
